+++ veltahleon/src/main/java/com/testament/veltahleon/abstraction/Human.java
@@ -9,6 +9,9 @@ import com.testament.veltahleon.model.entities.society.Title;
 import com.testament.veltahleon.model.entities.society.enumeration.Gender;
 
 import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
@@ -23,6 +26,9 @@ import java.util.List;
 @ToString
 public abstract class Human {
 
+    @NotNull(message = "Name cannot be null!")
+    @NotBlank(message = "Name cannot be blank!")
+    @NotEmpty(message = "Name cannot be empty!")
     @Column(name = "name")
     private String name;