+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/places/Province.java
@@ -2,6 +2,9 @@ package com.testament.veltahleon.model.entities.places;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
 @Entity
@@ -16,6 +19,9 @@ public class Province {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @NotNull(message = "Name cannot be null!")
+    @NotBlank(message = "Name cannot be blank!")
+    @NotEmpty(message = "Name cannot be empty!")
     @Column(name = "name")
     private String name;