+++ veltahleon/src/main/java/com/testament/veltahleon/repositories/places/CapitalRepository.java
@@ -9,4 +9,5 @@ public interface CapitalRepository extends JpaRepository<Capital, Long> {
 
     Capital findByName(String name);
     Capital findByNation_Name(String name);
+    long countByName(String name);
 }