+++ veltahleon/src/main/java/com/testament/veltahleon/services/implementation/places/CapitalServiceImpl.java
@@ -1,7 +1,10 @@
 package com.testament.veltahleon.services.implementation.places;
 
+import com.testament.veltahleon.exceptions.DataInsertionException;
 import com.testament.veltahleon.model.entities.places.Capital;
+import com.testament.veltahleon.model.entities.places.Nation;
 import com.testament.veltahleon.repositories.places.CapitalRepository;
+import com.testament.veltahleon.repositories.places.NationRepository;
 import com.testament.veltahleon.services.ifc.places.CapitalService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +23,8 @@ public class CapitalServiceImpl implements CapitalService {
 
     @Autowired
     private CapitalRepository capitalRepository;
+    @Autowired
+    private NationRepository nationRepository;
 
     @Override
     public Collection<Capital> getCapitalsWithPagination(int pageNumber, int numberOfRecords) {
@@ -54,6 +59,15 @@ public class CapitalServiceImpl implements CapitalService {
 
     @Override
     public Capital saveCapital(Capital capital) {
+        if(capitalRepository.countByName(capital.getName().toLowerCase()) >= 1) {
+            throw new DataInsertionException("Capital name");
+        }
+
+        if(nationRepository.countByName(capital.getNation().getName()) >= 1) {
+            Nation nation = nationRepository.findByName(capital.getNation().getName());
+            capital.getNation().setId(nation.getId());
+            capital.setNation(nation);
+        }
         return capitalRepository.save(capital);
     }
 
@@ -70,9 +84,46 @@ public class CapitalServiceImpl implements CapitalService {
         }
 
         if(capital.getNation() != null && newCapital.getNation() != capital.getNation()) {
-            newCapital.setNation(capital.getNation());
+            newCapital.setNation(checkNationForUpdate(capital.getNation(), newCapital.getNation()));
         }
 
         return capitalRepository.save(newCapital);
+    }
+
+    //Helper Methods
+    private Nation checkNationForUpdate(Nation nation, Nation newNation) {
+        if(nation.getName() != null && newNation.getName() != nation.getName()) {
+            newNation.setName(nation.getName());
+        }
+
+        if(nation.getDescription() != null && newNation.getDescription() != nation.getDescription()) {
+            newNation.setDescription(nation.getDescription());
+        }
+
+        if(nation.getCapital() != null && newNation.getCapital() != nation.getCapital()) {
+            newNation.setCapital(nation.getCapital());
+        }
+
+//        if(nation.getContinent() != null && newNation.getContinent() != nation.getContinent()) {
+//            newNation.setContinent(nation.getContinent());
+//        }
+
+        if(nation.getLeader() != null && newNation.getLeader() != nation.getLeader()) {
+            newNation.setLeader(nation.getLeader());
+        }
+
+        if(nation.getLanguage() != null && newNation.getLanguage() != nation.getLanguage()) {
+            newNation.setLanguage(nation.getLanguage());
+        }
+
+        if(nation.getProvinces() != null && newNation.getProvinces() != nation.getProvinces()) {
+            newNation.setProvinces(nation.getProvinces());
+        }
+
+        if(nation.getType() != null && newNation.getType() != nation.getType()) {
+            newNation.setType(nation.getType());
+        }
+
+        return newNation;
     }
 }